==> notebooks/05_preparation_des_donnees.ipynb <==
# Initialiser Spark

from pyspark import SparkContext, SparkConf
conf = SparkConf().setAppName("Notebook").setMaster("local")
sc = SparkContext(conf=conf)

# Vérifier le contenu du fichier

prenoms = sc.textFile("prenoms.txt")
prenoms.takeSample(False, 10)

# Extraire et sauvegarder un échantillon
1. Expliquer ce que fait la ligne suivante

prenoms.filter(lambda l: l.startswith("sexe") == False)\
    .filter(lambda l: "XX" not in l)\
    .sample(False, .001)\
    .saveAsTextFile("prenoms_sample.txt")

Cela retire les lignes avec des données manquante marqué par des XX.

# Vérifier l'échantillon

echantillon = sc.textFile("prenoms_sample.txt")
print(echantillon.count())
print(echantillon.filter(lambda l: l.startswith("sexe")).count())
print(echantillon.filter(lambda l: "XX" in l).count())
