==> notebooks/10_introduction_aux_rdd.ipynb <==
# Initialisation de Spark

from pyspark import SparkContext, SparkConf
conf = SparkConf().setAppName("Notebook").setMaster("local")
sc = SparkContext(conf=conf)

# Une fonction pour simplifier l'accès aux données

def extract_field(s: str, field_number: int, sep: str = ';') -> str:
    fields = s.split(sep)
    return fields[field_number] if (field_number < len(fields)) else ""

print(extract_field("2;CASSIOPEE;2009;33;3", 0))
print(extract_field("2;CASSIOPEE;2009;33;3", 1))
print(extract_field("2;CASSIOPEE;2009;33;3", 2))
print(extract_field("2;CASSIOPEE;2009;33;3", 3))
print(extract_field("2;CASSIOPEE;2009;33;3", 4))
print(extract_field("2;CASSIOPEE;2009;33;3", 5))

# Charger les données
1. Créer le RDD `lignes` à partir du répertoire `prenoms_sample.txt`

lignes = sc.textFile("prenoms_sample.txt")
print(lignes.take(10))

# Transformer les lignes en prénoms
1. En appliquant la méthode `map`, créer le RDD `prenoms` à partir de `lignes`

prenoms = lignes.map(lambda l: (
    extract_field(l, 0)[0],
    extract_field(l, 1),
    int(extract_field(l, 2)),
    extract_field(l, 3),
    int(extract_field(l, 4))
))
for n in prenoms.take(10):
    print(n)

# Interroger les données
La documentation des méthodes d'un RDD est disponible ([RDD](https://spark.apache.org/docs/latest/api/python/reference/api/pyspark.RDD.html)).

1. Rappeler ce que sont les *transformations* et les *actions*

En Spark, les *transformations* décrivent les opérations à faire sur les données sans les exécuter, tandis que les *actions* déclenchent réellement leur exécution pour produire un résultat.

2. Donner, pour chaque prénom, son nombre d'occurences (`map` et `reduceByKey`)

occ_prenoms = prenoms.map(lambda prenom: (prenom[1], 1)).reduceByKey(lambda a, b: a + b)

for n in occ_prenoms.take(10):
    print(n)

1. Donner le nombre total de naissances avec un prénom féminin (`filter`, `map`, `reduce` ou `sum`)

nb_naissance_feminin = prenoms.filter(lambda sexe: sexe[0] == '2').map(lambda prenom: prenom[4]).sum()
print(nb_naissance_feminin)

1. Donner l'effectif maximal et minimal par prénom (`map`, `aggregateByKey`)

max_min_prenoms = prenoms.map(lambda prenom: (prenom[1], prenom[4])).aggregateByKey(
    (float('inf'), float('-inf')),
    lambda acc, v: (min(acc[0], v), max(acc[1], v)),
    lambda acc1, acc2: (min(acc1[0], acc2[0]), max(acc1[1], acc2[1]))
    )

for n in max_min_prenoms.take(10):
    print(n)

1. Sur le modèle des prénoms, charger les données des départements
1. Donner, pour chaque nom de département, le prénom le plus fréquent depuis l'année 2000

dpts = sc.textFile("dpts.txt")
dpts.filter(lambda l: l.startswith("dep") == False)\
    .saveAsTextFile("dpts_sample.txt")


lignes = sc.textFile("dpts_sample.txt")

dpts = lignes.map(lambda l: (
    extract_field(l, 0, ','),
    int(extract_field(l, 1, ',')),
    (extract_field(l, 2, ',')),
    int(extract_field(l, 3, ',')),
    extract_field(l, 4, ','),
    extract_field(l, 5, ','),
    extract_field(l, 6, ',')
))
for n in dpts.take(10):
    print(n)

freq_depts_prenoms = prenoms.filter(lambda prenom: prenom[2] >= 2000)\
    .map(lambda prenom: ((prenom[1], prenom[3]), prenom[4]))\
    .reduceByKey(lambda a, b: a + b)
freq_by_dept = freq_depts_prenoms.map(lambda x: (x[0][1], (x[0][0], x[1])))  # clé = code_dept, valeur = (prenom, nb_naissances)

dpts_noms = dpts.map(lambda dpt: (dpt[0], dpt[4]))  # clé = code_dept, valeur = nom_dept

jointure = freq_by_dept.join(dpts_noms)\
    .reduceByKey(lambda a, b: a if a[0][1] > b[0][1] else b)\
    .map(lambda x: (x[1][1], x[1][0][0]))

for n in jointure.take(10):
   print(n)
