==> notebooks/30_spark_sql.ipynb <==
# Initialisation de Spark

from pyspark.sql import SparkSession
spark = SparkSession \
    .builder \
    .appName("Exemples avec les formats de stockage") \
    .getOrCreate()
sc = spark.sparkContext

# Chargement des données
1. Créer un dataframe pour les prénoms

prenoms = spark.read.load("prenomsParDeptsEtAnnees.gzip.parquet")
prenoms.createOrReplaceTempView("prenoms")
prenoms.show()

1. Faire de même pour les départements

departements = spark.read.load("depts.parquet")
departements.createOrReplaceTempView("departements")
departements.show()

# Interroger les données

* La documentation sur Spark SQL est disponible [ici](https://spark.apache.org/docs/latest/sql-programming-guide.html) (de même pour [DataFrame](https://spark.apache.org/docs/latest/api/python/reference/pyspark.sql/api/pyspark.sql.DataFrame.html))
* Pour chaque question, vous donnerez la réponse avec le DSL et en SQL

1. Donner, pour chaque prénom, son nombre d'occurences

## DSL

prenoms.groupBy("prenom").sum("effectif").show()

## SQL

spark.sql("""SELECT p.prenom, SUM(p.effectif)
             FROM prenoms p
             GROUP BY p.prenom""").show()

1. Donner le nombre total de naissances avec un prénom féminin

## DSL

prenoms.where("sexe = '2'").groupBy("sexe").sum("effectif").select("sum(effectif)").show()

## SQL

spark.sql("""SELECT SUM(effectif)
           FROM prenoms
           WHERE sexe = '2'""").show()

1. Donner l'effectif maximal et minimal par prénom

## DSL

from pyspark.sql.functions import min as _min, max as _max
prenoms.groupBy("prenom").agg(_min("effectif"),_max("effectif")).show()

## SQL

spark.sql("""SELECT p.prenom, MIN(effectif), MAX(effectif)
           FROM prenoms p
           GROUP BY p.prenom""").show()

1. Donner, pour chaque nom de département, le nombre moyen de prénoms depuis l'année 2000

## DSL

from pyspark.sql.functions import avg

prenoms.join(departements, prenoms.dep == departements.dep).filter("annee >= 2000").groupBy("ncc").avg("effectif").show()

## SQL

spark.sql("""SELECT d.ncc, AVG(p.effectif)
              FROM prenoms p, departements d
              WHERE p.dep = d.dep AND p.annee >= 2000
              GROUP BY d.ncc""").show()
